pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sksg-tchinese"
version = "1.1.0"
description = "Traditional Chinese localization tool for Hollow Knight: Silksong"
authors = []
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "UnityPy",
    "Pillow",
    "etcpak",
    "TypeTreeGeneratorAPI",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.0",
    "pytest-cov>=4.0",
]
build = [
    "pyinstaller",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
skip_covered = false
